Fixed package numbering
