+++ Level1-Module3/bin/.gitignore
@@ -2,7 +2,6 @@
 /_01_intro_to_static/
 /_01_static_smiley_blues/
 /_02_backpack/
-/_03_jars/
 /_05_netflix/
 /_06_gridworld/
 /_07_cow_timer/
@@ -10,3 +9,8 @@
 /gridworld.jar
 /javazoom.jar
 /league_token/
+/_02_static_smiley_blues/
+/_03_backpack/
+/_04_netflix/
+/_05_gridworld/
+/_06_cow_timer/
