--- Level1-Module3/src/_01_static_smiley_blues/SmileyBluesRunner.java
@@ -1,8 +1,0 @@
-package _01_static_smiley_blues;
-
-public class SmileyBluesRunner {
-	public static void main(String[] args) {
-		SmileyBlues smileyRunner = new SmileyBlues();
-		smileyRunner.start();
-	}
-}
