+++ Level1-Module3/src/_02_static_smiley_blues/SmileyBlues.java
@@ -1,0 +1,24 @@
+package _02_static_smiley_blues;
+
+public class SmileyBlues {
+	void start() {
+		// 1. Make a new JFrame and set it to be visible
+
+		// 2. Set your frame's default close operation to JFrame.EXIT_ON_CLOSE
+
+		// 3. Make a new JPanel
+
+		// 4. Add your panel to your frame
+
+		// 8. DO THIS LAST: 
+		// Set the Smiley class's color variable to Smiley.YELLOW (NOT a specific smiley object's!)
+		// Using a specific object for a static variable works, but is not usually recommended.
+		// Run it again; are your Smileys all yellow now? Also try changing them to red or green!
+
+		// 5. Make three Smiley objects and add them to your panel
+
+		// 6. Pack your frame
+
+		// 7. RUN THE PROGRAM and make sure you see three blue Smileys!
+	}
+}
