+++ Level1-Module3/src/_02_static_smiley_blues/SmileyBluesRunner.java
@@ -1,0 +1,8 @@
+package _02_static_smiley_blues;
+
+public class SmileyBluesRunner {
+	public static void main(String[] args) {
+		SmileyBlues smileyRunner = new SmileyBlues();
+		smileyRunner.start();
+	}
+}
