+++ Level1-Module3/src/_06_cow_timer/CowTimer.java
@@ -1,4 +1,4 @@
-package _07_cow_timer;
+package _06_cow_timer;
 /*
  *    Copyright (c) The League of Amazing Programmers 2013-2019
  *    Level 1
