+++ Level1-Module3/src/_06_cow_timer/CowTimerRunner.java
@@ -1,4 +1,4 @@
-package _07_cow_timer;
+package _06_cow_timer;
 
 public class CowTimerRunner {
 	public static void main(String[] args) throws InterruptedException {
