--- Level1-Module3/src/_06_gridworld/recipe.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html>
-<title>Recipes</title>
-<meta charset="UTF-8">
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<link rel="stylesheet" href="https://league-level0.github.io/style.css">
-<link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Raleway">
-
-<script src="https://league-level0.github.io/copyrightFooter.js"></script>
-<body>
-
-<script>
-window.onload=function(){document.location.href="http://central.jointheleague.org/levels/Level1/Mod3Recipes/gridworld.html"}
-</script>
-
- 
-
-<br><br><br><br><br><br><br><br><br>
-<div id="copyright">
-<script>
-copyright();
-</script>
-</div>
-	
-</body>
-</html>
